+++ package.json
@@ -20,8 +20,13 @@
   },
   "dependencies": {
     "expo": "^21.0.0",
+    "p-debounce": "^1.0.0",
     "react": "16.0.0-alpha.12",
     "react-native": "^0.48.4",
-    "react-native-elements": "^0.17.0"
+    "react-native-elements": "^0.17.0",
+    "react-navigation": "^1.0.0-beta.19",
+    "react-redux": "^5.0.6",
+    "redux": "^3.7.2",
+    "redux-thunk": "^2.2.0"
   }
 }
