+++ src/reducers/mainReducer.js
@@ -1,0 +1,12 @@
+let initialState = {
+  msg: 'hay'
+}
+
+export default (state = initialState, actions) => {
+  switch (actions.type) {
+    case "":
+      return state;
+    default:
+      return state;
+  }
+}
