+++ src/reducers/rootReducer.js
@@ -1,0 +1,9 @@
+import { combineReducers } from 'redux';
+
+import mainReducer from './mainReducer';
+
+const rootReducer = combineReducers({
+  mainStore: mainReducer
+})
+
+export default rootReducer;
