+++ src/util/tride.js
